evaluation/fakewebcam/requirements.txt
appdirs==1.4.3
blinker==1.4
click==6.6
dominate==2.2.0
Flask==0.11
Flask-Bootstrap==3.3.6.0
Flask-Mail==0.9.1
Flask-Moment==0.5.1
Flask-Redis==0.1.0
Flask-Script==2.0.5
Flask-SocketIO==2.8.5
Flask-SQLAlchemy==2.1
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mockredispy==2.9.3
nose==1.3.7
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.5.3
python-engineio==1.3.0
python-socketio==1.7.2
PyYAML==3.11
redis==2.10.5
requests==2.10.0
selenium==3.3.1
setproctitle==1.1.10
six==1.10.0
SQLAlchemy==1.0.13
visitor==0.1.3
Werkzeug==0.11.10
qrcode==5.3
